a corner case bug: when a chunk is downloaded. The has chunk file is updated, but somehow when requesting chunks the calculated # is wrong

+++ reliable_udp.h
@@ -27,7 +27,7 @@ typedef struct udp_sender_session {
     short last_acceptable_frame;
     short peer_id;
     char ip[IP_STR_LEN];
-    short sock;
+    int sock;
     char chunk_hash[CHUNK_HASH_SIZE];
     char *data;
     /* the size of stored data */
